Fix sanitation of submitted files. Closes #341.
Earlier, filenames such as "garbage-äß-.txt" were allowed, since we used
the \w character class on the call to gsub. However, this allows more
characters to be valid than we'd like to be. The reason for this is that
Ruby 1.8 when run in UTF-8 mode defines \w different from [0-9a-zA-Z].
This commit should fix this problem and tries to use the same regexp in
all places.