Closes #660. Rubric criterion failure.
The test made wrong assumptions about ordering. YAML files get
translated into hashes (which don't have a concept of order). The parsed
hash is then loaded into the DB and it so happened that the criteria
turned up in the wrong order. The fixed test no longer makes this
assumption.