Fix pagination of marks spreadsheet.
A couple of test errors were thrown and an exception showed up when the
"Grades" tab was selected for a marks spreadsheet. As it turnes out, the
problem was that the lambda took no arguments but was called with two
arguments. See app/controllers/grade_entry_forms_controller.rb
line 29. This fix only changes the offending lambda to take two dummy
arguments. That is needed as the pagination helper routine is also
called by the submissions table controllers.