Migrate .rvmrc to .ruby-version/.ruby-gemset files
.ruby-version/.ruby-gemset are compatible across the other popular Ruby 
environment managers and has been supported in RVM for over 2 years.

Additionally, RVM will automatically setup the gemset, including the
setup/installation of Bundler.

However, it will only show the command to install a Ruby version, it
will not install it automatically. It will also not use bundler to
install everything in the Gemfile.