Exclude inactive students from assignment stats
Inactive students should not be included when calculating the mean and
the median of the students' marks for an assignment.

This also improves the performance of the method, as it no longer
queries for a list of ActiveRecords (without eagerly loading the
associations as the current code does) objects but only a list of
integers.