Remove unmeaningful assertions
The assertions depend on the return value of `(un)?hide_students`, which
is not documented nor meaningful (update always return the found objects).

It is also failing since the last line of the two methods are no longer
a `Student.update(...)`, but instead of fixing the implementation, the
unmeaningful assertions can simply be removed. Exceptions raised in
tests count as test failure.