Fixed buggy code in modesl/user.rb
Frontend input fields don't allow characters like newlines and null-terminators and hence requests would never really contain these characters.

However, an attacker can easily bypass this using a program like curl and craft a request sending a newline character in the either the username or password field. This will cause the app to produce an error and can produce a back-trace. Implications of this should be obvious.