Catch exceptions when populating submissions table
Before, when any submission got into an invalid state, the helper
function gathering all the table data would fail. This caused the table
to not load at all. Now the error is caught, and an error boolean is
placed in the returned object, which the front-end then checks and
displays an errored submission (a light red row) in the table.
Population of other submissions should continue unaffected.