We saw a problem where the grace period was apparently not calculated correctly and the wrong revision was collected.
The logs showed that the instructor collected all the submissions more than
once. Ultimately the problem was that rows were being deleted from the
grace_period_deduction table, but the cached object didn't change when
computing the available grace credits.

The fix is to reload the grace_period_deductions so that the available
grace credits is correctly computed.