BUG FIX: Existing bug after removing image annot.
While implementing the delete item for the annotation context menu, an
existing bug was revealed:

Bug description:
After deleting an annotation that was on an image file, viewing the
image file again without refreshing would cause many
```
Uncaught TypeError: Cannot read property 'style' of null
    at ImageAnnotationGrid.draw_holders (ImageAnnotationGrid.js:83)
```
errors since the annotation was not properly removed from the grid.

Problem and solution:
The code for comparing the annotation being removed, was comparing
string === int. The solution was simple - convert one the int to string
before comparing.