BUG FIX: Code Annot. not showing after delete
Bug description:
Pre-existing bug that occurred after deleting an existing annotation.
After deleting an existing code/text annotation, the remaining
annotations do not show on-hover.

Bug resolution:
Evident from the changes in this commit, there were 2 things that caused
this bug. First, when calling splice, the second argument should be the
number of items to remove which is 1. It should not have been the
existing object. The second issue was that we were assignment the
relationships variable the result of the splice, which would just
returned the removed item. What we want is the resulting list after
splicing.