BUGFIX (pre-existing): Common annot. and images
Bug description:
It seems that the common/pre-canned annotations were not working for
images before any of the context-menu changes were introduced.

Bug solution:
In the annotations_controller, the `add_existing_annotation` endpoint
was not passing all the necessary information in the case of image
annotations. Properties such as `annotation_text_id` and `creator_id`
were missing. This has now been added, and we are now using the banged
version for `update_attributes!` for images (to stay consistent with how
pdf and text).