Codeviewer scrollbar fix (see description)
After making use of flex boxes with the panes, the expected behaviour of
the scrollbars for the codeviewer was changed. The horizontal scrollbar
was only accessible after scrolling all the way down vertically. Also,
scrollbars were not in the expected location of the codeviewer did not
have enough code to fill up the pane. This commit fixes these issues.