Change the VM from Xenial64 Trusty64
The reason for this change is because no solution for setting up the VM
using Xenial64 could be found within a reasonable amount of time when
using VirtualBox. The reason why we are now using VirtualBox as the
provider is because on Windows, Hyper-V is not available on the Home
edition.