Update the provisioning script so that it is split up
The provisioning script is now split into the configuration, pre-setup,
and setup steps. The order to execute the scripts in is pre-setup, then
setup, and finally the configuration. This is done using "vagrant
provision --provision-with x" where "x" is the name of the script to be
executed. At this point the script doesn't work because there still
remains an issue with yarn. Also, in the future, the various "echo"
commands which update the user of progress will be changed.