Upgrade to ruby 2.5.0 and nokogiri 1.8.2
It may not have been necessary to upgrade to nokogiri 1.8.2 because for
some reason using ruby 2.5.0 with nokogiri 1.8.1 did not cause any
issues, but because that issue had come up in the past, nokogiri has
been upgraded just to be safe.

The --disable-binary option when installing ruby was used because an
error was caused when it was not used, details can be found at
https://github.com/rvm/rvm/issues/4291

The reordering solved a problem with webpacker, this may have been
caused in an earlier commit but not caught, either way, that is why it
exists here. The issue was that project-specific dependencies and
webpacker adding was being done without first cloning and switching to
the Markus folder.