Mark Spreadsheets, Student Viewable Messages (#3367)
* Add a div for the message for students in the mark spreadsheets

Add the div under the condition that a message has been provided. Also
including styling.

* Update logic for message in student mark spreadsheet

Previously the block of code that had the message was only appearing if
marks had been released, this has been changed so that it will appear
regardless so long as there is a message.

* Add placeholder message to edit mark spreadsheet page

The message is currently not initialized as a constant, this will be
done in future commits.

* Edit the CSS for the grade message such that it uses constants

Previously it was not using constants, now it is. Also, some additional
formatting in the student_interface.html.erb was done but it was minor.

* Style changes to CSS properties of marks_message

First, the properties were reordered to alphabetical order.
Second, the integers were changed into their decimal equivalents.
Third, the shorthand was optimized.
Fourth, the 0em was re-written to 0.

* More CSS style changes for marks_message

The 1.0em from the previous change has been rewritten as 1em.

* Rename div class of marks_message to marks-message

* Add constant in GradeEntryForm model for Mark Spreadsheet Message

The constant is added in the GradeEntryForm model and called in the view
of the form.

* Style changes to MESSAGE_PLACEHOLDER in GradeEntryForm model

* Freeze the constant MESSAGE_PLACEHOLDER in module GradeEntryForm

* Switch the constant usage of MESSAGE_PLACEHOLDER to be internationalized

The >- usage in the .yml file does not seem to be super common, but it
was done to keep the character line limit down.