submissions: Enable assignment only_required_files with subdirectories.
Previously, students were able to submit files into subdirectories
whether or not these files were outlawed by the assignment's required
file list and "only_required_files" setting.

This was slightly tricky because the add_files method took in a relative
path to the repo root (including the assignment folder), while the
required_files list contained paths only relative to the assignment folder.
This commit changes the latter, so now the required_files list are
relative to the repo root.

There was also a slight bug with how we were handling uploading new
files vs. replacing existing files: the latter was not being guarded
against by the only_required_files setting. This commit fixes that as
well.