assignments: handle remarks in get_num_marked.
This fixes Assignment#get_num_marked so that it correctly counts
groupings with remark requests. Previously, when a grouping submitted a
remark request, both the original and remark result would be counted,
rather than just the remark result. The cause was the same error as
noted in #5063: joining groupings on :current_result doesn't limit to
one result per grouping.

The root fix is the creation of a new method Assignment#current_results,
which uses a more sophisticated db query to retrive one current result
per grouping.

This commit also simplifies the code for calculating the number of
marked submissions for a grader when graders are assigned to criteria.