Fix CC license violation (#6441)
The web interface is currently using icons by Mark James from his "silk" icon set: https://github.com/markjames/famfamfam-silk-icons

These icons are licensed under the Creative Commons Attribution 2.5 Generic (CC BY 2.5), which requires attribution.

To the best of my knowledge, there is no notice of attribution in the project for the use of these icons. Apologies if attribution exists and I missed it.

There are 2 possible solutions to this license violation:
1. Remove the silk icons from the project.
2. Add a notice to the README which discloses the use of the icons.

This PR opts for the second solution and adds a notice to the end of README.md.