enriching-data-returned-by-graders-controller-index-action
Our goal is to be able to determine in the ffront end which groups should be hidden because they are inactive. Our definition of an inactive group is that all of its members are inactive. Thus, (along with the graders associated to each group, which we already passing to the front end) we need to pass the (student) members associated with each group. This will allow the front end for this graders use case to act similarly to the front end of the groups use case (by filtering which groups should be visible or not based on the activity/inacticity of their users)