prventing-assigning-of-inactive-graders
Implementing the following functionality: When an instructor tries to add an inactive grader to a group, the system prevents him from doing so and throws a red flash-message saying the user is inactive and thus cannot be added. NOTE: If some graders are active and some are inactive, then the active graders will be added, while the inactive will not be added.