tests-and-small-refactoring
Wrote five tests for assigning, random-assigning, and un-assigning of graders (to groupings) by instructors, ensuring the expected behavior is exhibited (for instance not allowing inactive graders to be assigned to a grouping, but allowing their removal). Some new Ta objects were created for this purpose in the tests. Additionally, this commit includes some refactorings of the graders_controller:global_actions action. For example, I removed a redundant 'begin-rescue' block (which I added in first pull request for the bug, where I needed it to raise the flash message).