section-destroy-api-route
Added a new api route for 'deleting' a section. This included creating a new sections_controller in the api sub-package, and specifying 'resources' line to the ~/config/routes.rb file (in the appropriate location so that the new api URI's look like .../courses/.../sections/...). I also created a new rspec_test testing the functionality of this new API route, e.g. for authorized vs unauthorized attempts.