comments-and-style
Firstly, I commented on the line ‘grader_ids.map!(&:to_i)’ in the graders_controller to explain why it is necessary. Also, in the tests in ‘graders_controller_spec.rb’, in all requests which specify parameter ‘groupings’ or ‘graders’, I applied the style ‘groupings: [<grouping>.id,…]’ and ‘graders: [<grader>.id,…]’ for consistency and explicitness.