actions-for-sections-api-controller
Added index, show, and update actions for the api controller of the sections model. I modified routes.rb (to actually establish the existence of these routes).

I realised that because of the MainApiController (which all the API controllers inherit), the SessionHandler.check_record method already ensures that an incoming request trying to do something with a particular record is valid (i.e. the record actually exists), so it would be redundant to check whether current_course.sections.find(params[:id]) for example returns a valid current_course.