tests-for-section-api-controller
Added rspec tests for the update, index, and show actions of the api controller of sections. I tested both authorised and unauthorised requests, and for the GET requests (show and index), where either a json or xml response is expected, I tested that all necessary keys (corresponding to model attributes) are present, although I did not check whether the values of these attributes match between the responses and the original object data (not sure whether I should go this far). Also, for the #index request, some of my expectations perform an equality comparison between arrays; maybe it would be preferable to instead convert these to sets and make a comparison between sets.