fix-assignment-destroy-bug
Due to the nesting of button_to into form_for, in the generated HTML, Rails included two hidden _method attributes -the first being 'put' for form_for and the second being 'delete' for button_to-, and (by default behaviours) priority was given to the second definition, so when trying to update an assignment, Rails automatically tried to destroy it! Nesting the button_to helper into the form_for helper is by definition wrong, since HTML does not allow for nested forms (and both of these helpers generate forms), so the behavior of doing this is undefined. To fix, I took the button_to declaration out, so it is no longer nested inside form_for.