Delete functionality for peer-review assessments.
Turns out that, when editing an assessment, a different partial (peer_review_form) is rendered if the assessment is of type peer-review, so I added the new delete button (currently found only in the _edit partial) to that partial as well. Not sure if it is worth it to extract this button definition into a different partial and reuse it.