Fix problem with missing flash message
In the context of deleting an assignment, the green flash for informing the user that the assignment was successfully deleted works as expected except in one scenario: if the assignment was just created. That is, if you create a new assignment adn then on the spot (i.e. without leaving the edit page) try to delete, it will delete successfuly but the green flash message will not show. I think this has something to do with how flash messages work in Rails deep down, but could not figure exactly the cause. The added line removes the :success key of the flash so it can start fresh with the flash responder.