Fix Bug in Updating Assignments (#6962)
* fix-assignment-destroy-bug

Due to the nesting of button_to into form_for, in the generated HTML, Rails included two hidden _method attributes -the first being 'put' for form_for and the second being 'delete' for button_to-, and (by default behaviours) priority was given to the second definition, so when trying to update an assignment, Rails automatically tried to destroy it! Nesting the button_to helper into the form_for helper is by definition wrong, since HTML does not allow for nested forms (and both of these helpers generate forms), so the behavior of doing this is undefined. To fix, I took the button_to declaration out, so it is no longer nested inside form_for.

* Delete functionality for peer-review assessments.

Turns out that, when editing an assessment, a different partial (peer_review_form) is rendered if the assessment is of type peer-review, so I added the new delete button (currently found only in the _edit partial) to that partial as well. Not sure if it is worth it to extract this button definition into a different partial and reuse it.

* Fix problem with missing flash message

In the context of deleting an assignment, the green flash for informing the user that the assignment was successfully deleted works as expected except in one scenario: if the assignment was just created. That is, if you create a new assignment adn then on the spot (i.e. without leaving the edit page) try to delete, it will delete successfuly but the green flash message will not show. I think this has something to do with how flash messages work in Rails deep down, but could not figure exactly the cause. The added line removes the :success key of the flash so it can start fresh with the flash responder.