Handle case where no test groups are runnable by students
In the autotest-settings page (in the instructor view), the instructor has the ability to make all test groups NOT-runnable by students. Currently, a request is still sent to the autotesting server and 0/0 is shown on the website. Instead, we check this inside the automated_tests_controller#execute_test_run action, and if NO test groups are student-runnable, we flash an appropriate messgae and return, avoiding a redundant request to the autotesting server.