Testing for the change
I had to modify the current tests because the change of this pull request does change the behavior of automated testing. In particular, if none of the test-groups are runnable by students, now the test-execution job does not get enqeued as a job (which was the whole point, since if it gets enqueued to AutotestRunJob then an unnecessary request will be made to the autotest server). But this made some of the existing tests for execute_test_run (in automated_test_controller.rb) fail. This is because the factory-produced assignment object used in these tests had the test_groups attribute/association uninitialized. To test the new behavior, instead of generating a simple :assignment, the tests now use an :assignment_with_criteria_and_test_results, and in the tests with context , I had to manually update the test_groups attribute of the assignment to specify the test_group is runnable by students. Overall, we are now testing the case where none of the test groups are runnable by students, so no job should be enqueued, and the case where at least one test-group is student-runnable, in which case the job should be enqueued as usual.