association-improvement (#6989)
Added dependent: :destroy to all the associations that needed it (in the Assignment and Assessment models), changed the order of associations in Assignment to make groups-and-groupings first (this is the first thing  that should be checked when deleting an assignment), created a new locale string for use in the AnnotationCategory model, and added another rescue block in the #destroy action of assignments controller to handle other potential complications that could arise in assignment deletion. This last one is temporary, since we would ideally like to perform a more targeted handling of different exceptions (but for now this prevents crashing and gives an idea of the problem.