Handle case where no test groups are runnable by students (#7003)
* Handle case where no test groups are runnable by students

In the autotest-settings page (in the instructor view), the instructor has the ability to make all test groups NOT-runnable by students. Currently, a request is still sent to the autotesting server and 0/0 is shown on the website. Instead, we check this inside the automated_tests_controller#execute_test_run action, and if NO test groups are student-runnable, we flash an appropriate message and return, avoiding a redundant request to the autotesting server.

* Testing for the change

I had to modify the current tests because the change of this pull request does change the behavior of automated testing. In particular, if none of the test-groups are runnable by students, now the test-execution job does not get enqeued as a job (which was the whole point, since if it gets enqueued to AutotestRunJob then an unnecessary request will be made to the autotest server). But this made some of the existing tests for execute_test_run (in automated_test_controller.rb) fail. This is because the factory-produced assignment object used in these tests had the test_groups attribute/association uninitialized. To test the new behavior, instead of generating a simple :assignment, the tests now use an :assignment_with_criteria_and_test_results, and in the tests with context , I had to manually update the test_groups attribute of the assignment to specify the test_group is runnable by students. Overall, we are now testing the case where none of the test groups are runnable by students, so no job should be enqueued, and the case where at least one test-group is student-runnable, in which case the job should be enqueued as usual.

* Update Changelog

* Testing and querying improvements

Refactor the new testing by creating new factories for assignments with/without student-runnable tests (and, by extension, corresponding test-group factories) and use  to improve query efficiency.

* minor improvements

Improve test-group retrieval using pluck (both active record pluck and enumerable pluck) and remove redundant declarations from the rspec tests

---------

Co-authored-by: David Liu <david@cs.toronto.edu>